+++ django_enumfield/enum.py
@@ -1,93 +1,124 @@
+from __future__ import absolute_import
+
 import logging
+from enum import Enum as NativeEnum, IntEnum as NativeIntEnum
+from typing import Any, Dict, List, Optional, Sequence, Tuple, TypeVar, Union, cast
 
 from django.utils import six
-from django.utils.encoding import python_2_unicode_compatible
+from django.utils.decorators import classproperty
 
 from django_enumfield.db.fields import EnumField
 
-
 logger = logging.getLogger(__name__)
+RAISE = object()
 
 
-class EnumType(type):
-    """ Custom metaclass for Enum type """
+class BlankEnum(NativeEnum):
+    BLANK = ""
 
-    def __new__(mcs, *args):
-        """ Create enum values from all uppercase class attributes and store them in a dict on the Enum class."""
-        enum = super(EnumType, mcs).__new__(mcs, *args)
-        attributes = [k_v for k_v in list(enum.__dict__.items()) if k_v[0].isupper()]
-        labels = enum.__dict__.get('labels', {})
-
-        enum.values = {}
-        for attribute in attributes:
-            enum.values[attribute[1]] = enum.Value(attribute[0], attribute[1], labels.get(attribute[1]), enum)
-        return enum
+    @property
+    def label(self):
+        return ""
 
 
-class Enum(six.with_metaclass(EnumType)):
+def classdispatcher(class_method):
+    class _classdispatcher(object):
+        def __init__(self, method=None):
+            self.fget = method
+
+        def __get__(self, instance, cls=None):
+            if instance is None:
+                return getattr(cls, class_method)
+            return self.fget(instance)
+
+    return _classdispatcher
+
+
+Default = TypeVar("Default")
+T = TypeVar("T", bound="Enum")
+
+
+class Enum(NativeIntEnum):
     """ A container for holding and restoring enum values """
 
-    @python_2_unicode_compatible
-    class Value(object):
+    __labels__ = {}  # type: Dict[int, six.text_type]
+    __default__ = None  # type: Optional[int]
+    __transitions__ = {}  # type: Dict[int, Sequence[int]]
+
+    @classdispatcher("get_name")
+    def name(self):
+        # type: () -> str
+        return self._name_
+
+    @classdispatcher("get_label")
+    def label(self):
+        # type: () -> str
+        """ Get human readable label for the matching Enum.Value.
+        :return: label for value
+        :rtype: str
         """
-        A value represents a key-value pair with a uppercase name and a integer value:
-        GENDER = 1
-        "name" is a upper case string representing the class attribute
-        "label" is a translatable human readable version of "name"
-        "enum_type" is the value defined for the class attribute
+        label = cast(str, self.__class__.__labels__.get(self.value, self.name))
+        return six.text_type(label)
+
+    @classproperty
+    def do_not_call_in_templates(cls):
+        # type: () -> bool
+        # Fix for Django templates so that any lookups of enums won't fail
+        # More info: https://stackoverflow.com/questions/35953132/how-to-access-enum-types-in-django-templates  # noqa: E501
+        return True
+
+    @classproperty
+    def values(cls):  # type: ignore
+        return {member.value: member for member in cls}
+
+    def deconstruct(self):
         """
+        See "Adding a deconstruct() method" in
+        https://docs.djangoproject.com/en/1.8/topics/migrations/
+        """
+        c = self.__class__
+        path = "{}.{}".format(c.__module__, c.__name__)
+        return path, [self.value], {}
 
-        def __init__(self, name, value, label, enum_type):
-            self.name = name
-            self.value = value
-            self._label = label
-            self.enum_type = enum_type
-
-        def __str__(self):
-            return six.text_type(self.label)
-
-        def __repr__(self):
-            return self.name
-
-        def __eq__(self, other):
-            if other and isinstance(other, Enum.Value):
-                return self.value == other.value
-            elif isinstance(other, six.string_types):
-                return type(other)(self.value) == other
-            else:
-                raise TypeError('Can not compare Enum with %s' % other.__class__.__name__)
-
-        @property
-        def label(self):
-            return self._label or self.name
-
-        def deconstruct(self):
-            path = self.__module__ + '.' + self.__class__.__name__
-            return path, (self.name, self.value, self.label, self.enum_type), {}
+    @classmethod
+    def items(cls):
+        # type: () -> List[Tuple[str, int]]
+        """
+        :return: List of tuples consisting of every enum value in the form
+            [('NAME', value), ...]
+        """
+        items = [(member.name, member.value) for member in cls]
+        return sorted(items, key=lambda x: x[1])
 
     @classmethod
     def choices(cls, blank=False):
+        # type: (bool) -> List[Tuple[Union[int, str], NativeEnum]]
         """ Choices for Enum
-        :return: List of tuples (<value>, <human-readable value>)
-        :rtype: list
+        :return: List of tuples (<value>, <member>)
         """
-        choices = sorted([(key, value) for key, value in cls.values.items()], key=lambda x: x[0])
+        choices = sorted(
+            [(member.value, member) for member in cls], key=lambda x: x[0]
+        )  # type: List[Tuple[Union[str, int], NativeEnum]]
         if blank:
-            choices.insert(0, ('', Enum.Value('', None, '', cls)))
+            choices.insert(0, (BlankEnum.BLANK.value, BlankEnum.BLANK))
         return choices
 
     @classmethod
     def default(cls):
-        """ Default Enum value. Override this method if you need another default value.
+        # type: () -> Optional[Enum]
+        """ Default Enum value. Set default value to `__default__` attribute
+        of your enum class or override this method if you need another
+        default value.
         Usage:
             IntegerField(choices=my_enum.choices(), default=my_enum.default(), ...
-        :return Default value, which is the first one by default.
-        :rtype: int
+        :return Default value, if set.
         """
-        return cls.choices()[0][0]
+        if cls.__default__ is not None:
+            return cast(Enum, cls(cls.__default__))
 
     @classmethod
     def field(cls, **kwargs):
+        # type: (Any) -> EnumField
         """ A shortcut for field declaration
         Usage:
             class MyModelStatuses(Enum):
@@ -102,69 +133,84 @@ class Enum(six.with_metaclass(EnumType)):
         return EnumField(cls, **kwargs)
 
     @classmethod
-    def get(cls, name_or_numeric):
+    def get(
+        cls,
+        name_or_numeric,  # type: Union[str, int, T]
+        default=None,  # type: Optional[Default]
+    ):
+        # type: (...) -> Union[Enum, Optional[Default]]
         """ Get Enum.Value object matching the value argument.
         :param name_or_numeric: Integer value or attribute name
-        :type name_or_numeric: int or str
-        :rtype: Enum.Value
+        :param default: The default to return if the value passed is not
+            a valid enum value
         """
-        if isinstance(name_or_numeric, six.string_types):
-            name_or_numeric = getattr(cls, name_or_numeric.upper())
+        if isinstance(name_or_numeric, cls):
+            return name_or_numeric
 
-        return cls.values.get(name_or_numeric)
+        if isinstance(name_or_numeric, int):
+            try:
+                return cls(name_or_numeric)
+            except ValueError:
+                pass
+        elif isinstance(name_or_numeric, six.string_types):
+            try:
+                return cls[name_or_numeric]
+            except KeyError:
+                pass
 
-    @classmethod
-    def name(cls, numeric):
-        """ Get attribute name for the matching Enum.Value
-        :param numeric: Enum value
-        :type numeric: int
-        :return: Attribute name for value
-        :rtype: str
-        """
-        return cls.get(numeric).name
-
-    @classmethod
-    def label(cls, numeric):
-        """ Get human readable label for the matching Enum.Value.
-        :param numeric: Enum value
-        :type numeric: int
-        :return: label for value
-        :rtype: str or
-        """
-        return six.text_type(cls.get(numeric).label)
+        return default
 
     @classmethod
-    def items(cls):
+    def get_name(cls, name_or_numeric):
+        # type: (Union[str, int, T]) -> Optional[str]
+        """ Get Enum.Value name matching the value argument.
+        :param name_or_numeric: Integer value or attribute name
+        :return: The name or None if not found
         """
-        :return: List of tuples consisting of every enum value in the form [('NAME', value), ...]
-        :rtype: list
+        value = cls.get(name_or_numeric)
+        if value is not None:
+            return value.name
+        return None
+
+    @classmethod
+    def get_label(cls, name_or_numeric):
+        # type: (Union[str, int, Enum]) -> Optional[str]
+        """ Get Enum.Value label matching the value argument.
+        :param name_or_numeric: Integer value or attribute name
+        :return: The label or None if not found
         """
-        items = [(value.name, key) for key, value in cls.values.items()]
-        return sorted(items, key=lambda x: x[1])
+        value = cls.get(name_or_numeric)
+        if value is not None:
+            return value.label
+        return None
 
     @classmethod
     def is_valid_transition(cls, from_value, to_value):
-        """ Will check if to_value is a valid transition from from_value. Returns true if it is a valid transition.
+        # type: (Union[int, Enum], Union[int, Enum]) -> bool
+        """ Will check if to_value is a valid transition from from_value.
+        Returns true if it is a valid transition.
+
         :param from_value: Start transition point
         :param to_value: End transition point
-        :type from_value: int
-        :type to_value: int
         :return: Success flag
-        :rtype: bool
         """
-        try:
-            return from_value == to_value or from_value in cls.transition_origins(to_value)
-        except KeyError:
-            return False
+        if isinstance(from_value, cls):
+            from_value = from_value.value
+        if isinstance(to_value, cls):
+            to_value = to_value.value
+
+        return (
+            from_value == to_value
+            or not cls.__transitions__
+            or (from_value in cls.transition_origins(to_value))
+        )
 
     @classmethod
     def transition_origins(cls, to_value):
+        # type: (Union[int, T]) -> Sequence[int]
         """ Returns all values the to_value can make a transition from.
         :param to_value End transition point
-        :type to_value: int
-        :rtype: list
         """
-        return cls._transitions[to_value]
-
-
-Value = Enum.Value
+        if isinstance(to_value, cls):
+            to_value = to_value.value
+        return cast(Sequence[int], cls.__transitions__.get(to_value, []))