+++ django_enumfield/tests/models.py
@@ -9,9 +9,11 @@ class LampState(Enum):
     OFF = 0
     ON = 1
 
+    __default__ = OFF
+
 
 class Lamp(models.Model):
-    state = EnumField(LampState)
+    state = EnumField(LampState, verbose_name="stately_state")
 
 
 class PersonStatus(Enum):
@@ -21,21 +23,31 @@ class PersonStatus(Enum):
     REANIMATED = 3
     VOID = 4
 
-    _transitions = {
+    __transitions__ = {
         UNBORN: (VOID,),
         ALIVE: (UNBORN,),
         DEAD: (UNBORN, ALIVE),
-        REANIMATED: (DEAD,)
+        REANIMATED: (DEAD,),
     }
 
 
+class PersonStatusDefault(Enum):
+    UNBORN = 0
+    ALIVE = 1
+    DEAD = 2
+    REANIMATED = 3
+    VOID = 4
+
+    __default__ = UNBORN
+
+
 class Person(models.Model):
-    example = models.CharField(max_length=100, default='foo')
+    example = models.CharField(max_length=100, default="foo")
     status = EnumField(PersonStatus, default=PersonStatus.ALIVE)
 
     def save(self, *args, **kwargs):
         super(Person, self).save(*args, **kwargs)
-        return 'Person.save'
+        return "Person.save"
 
 
 class BeerStyle(Enum):
@@ -43,16 +55,15 @@ class BeerStyle(Enum):
     STOUT = 1
     WEISSBIER = 2
 
+    __default__ = LAGER
+
 
 class BeerState(Enum):
     FIZZY = 0
     STALE = 1
     EMPTY = 2
 
-
-class Beer(models.Model):
-    style = EnumField(BeerStyle)
-    state = EnumField(BeerState, null=True, blank=True)
+    __default__ = FIZZY
 
 
 class LabelBeer(Enum):
@@ -60,7 +71,14 @@ class LabelBeer(Enum):
     JUPILER = 1
     TYSKIE = 2
 
-    labels = {
-        STELLA: _('Stella Artois'),
-        TYSKIE: _('Browar Tyskie'),
-    }
+    __labels__ = {STELLA: _("Stella Artois"), TYSKIE: _("Browar Tyskie")}
+
+
+def get_default_beer_label():
+    return LabelBeer.JUPILER
+
+
+class Beer(models.Model):
+    style = EnumField(BeerStyle)
+    state = EnumField(BeerState, null=True, blank=True)
+    label = EnumField(LabelBeer, default=get_default_beer_label)